Created nodelogin directory / Updated README files
Created nodelogin directory to implement a basic login and registration system that runs off of a local database. We need to implement postgresql into this directory. I have also updated each of the directory's README files to give an understanding of what they are utilized for and how to get them working on your machine.