Updated toolbar pages / Updated database with Danial's queries
Updated the toolbar pages with axios fetch / get commands. I also updated the database to include Danial's contributions. So now -- the database is full of the tables we need. I implemented basic axios / express commands to fetch this information. Fixing the JSON(s) displaying is the most important and next step to complete this project(!)